make build script more understandable

+++ build.musl.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 set -exo pipefail
-# build in the currently available muslrust container
-container="$(docker images -q edonusdevelopers/muslrust | head -n 1)"
 
 cargo_cache() {
   # Allow using a build cache when not on the build workers
@@ -23,32 +21,26 @@ docker_run() {
   docker run --rm $cache -v "$PWD:/volume" -w /volume -t "${container}" $@
 }
 
-# compile test executable
-docker_run cargo build --test testmain
+run_tests() {
+  # compile test executable
+  docker_run cargo build --test testmain
 
-# ensure we don't overwrite a buildmachines config
-# back it up, then restore on EXIT
-[ -f ~/.lal/config ] && cp ~/.lal/config ./bupconfig
-restore_config() {
-  [ -f bupconfig ] && mv bupconfig ~/.lal/config
+  # ensure we don't overwrite a buildmachines config
+  # back it up, then restore on EXIT
+  [ -f ~/.lal/config ] && cp ~/.lal/config ./bupconfig
+  restore_config() {
+    [ -f bupconfig ] && mv bupconfig ~/.lal/config
+  }
+  trap restore_config EXIT
+
+  # run tests
+  ./target/x86_64-unknown-linux-musl/debug/testmain-*
+  rm -rf ARTIFACT
 }
-trap restore_config EXIT
 
-# run tests
-./target/x86_64-unknown-linux-musl/debug/testmain-*
 
-# compile lal
-docker_run cargo build --release --verbose
-
-# create release tarball in right dir structure for artifactory
-lalversion=$(grep version Cargo.toml | awk -F"\"" '{print $2}')
-
-buildurl="http://engci-maven.cisco.com/artifactory/api/storage/CME-release/lal"
-if curl -s "${buildurl}" | grep -q "$lalversion"; then
-    echo "lal version already uploaded - stopping" # don't want to overwrite
-    rm -rf ARTIFACT # tests create a release build..
-else
-  echo "Packaging new lal version"
+build_lal_tarball() {
+  docker_run cargo build --release --verbose
   mkdir -p musl/bin
   mkdir -p musl/share/lal/configs
   cp target/x86_64-unknown-linux-musl/release/lal musl/bin
@@ -56,9 +48,33 @@ else
   cp configs/* musl/share/lal/configs/
   tar czf lal.tar -C musl .
   rm -rf musl/
-  rm -rf ARTIFACT
-  mkdir "ARTIFACT/${lalversion}" -p
-  cp lal.tar "ARTIFACT/${lalversion}/"
-  # Update the latest package
-  cp "ARTIFACT/${lalversion}" "ARTIFACT/latest" -R
-fi
+  echo "Created lal tarball with contents:"
+  tar tvf lal.tar
+}
+
+create_lal_upload() {
+  # Upload to a folder on artifactory equal to the Cargo.toml version
+  lalversion=$(grep version Cargo.toml | awk -F"\"" '{print $2}')
+  # But only if that folder doesn't already exist
+  buildurl="http://engci-maven.cisco.com/artifactory/api/storage/CME-release/lal"
+  if curl -s "${buildurl}" | grep -q "$lalversion"; then
+      echo "lal version already uploaded - stopping" # don't want to overwrite
+  else
+    echo "Packaging new lal version"
+    mkdir "ARTIFACT/${lalversion}" -p
+    cp lal.tar "ARTIFACT/${lalversion}/"
+    # Update the latest package
+    cp "ARTIFACT/${lalversion}" "ARTIFACT/latest" -R
+  fi
+}
+
+main() {
+  # build in the currently available muslrust container
+  local -r container="$(docker images -q edonusdevelopers/muslrust | head -n 1)"
+  run_tests
+  build_lal_tarball
+  create_lal_upload
+  rm lal.tar
+}
+
+main
